♻️ refactor(frontend/queryclient): move qureyclient and router to App (re #43)
Because TanStack's Query recommands not to use QueryClient with export-import. Not affecting current
app's logic, just for future updates.

re #43
