+++ packages/client/app/src/RouteGuard/components/CheckFtOAuth.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { useAtomValue } from 'jotai';
 import { Navigate } from 'react-router-dom';
-import { RouteList } from '../../AppRouter';
+import { RouteList } from '../../App';
 import { needFtOAuthAtom } from '../../Login/atoms/needFtOAuthAtom';
 import { JSXChildren } from '../../types/JSXChildren';
 import { useSetOriginalPath } from '../hooks/useSetOriginalPath';
