+++ packages/client/app/src/RouteGuard/components/CheckLogin.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { useAtomValue } from 'jotai';
 import { Navigate } from 'react-router-dom';
-import { RouteList } from '../../AppRouter';
+import { RouteList } from '../../App';
 import { needLoginAtom } from '../../Login/atoms/refreshTokenAtom';
 import { JSXChildren } from '../../types/JSXChildren';
 import { useSetOriginalPath } from '../hooks/useSetOriginalPath';
