+++ packages/client/app/src/RouteGuard/components/CheckNoFtOAuth.tsx
@@ -1,7 +1,7 @@
 import { useAtomValue } from 'jotai';
 import { needFtOAuthAtom } from '../../Login/atoms/needFtOAuthAtom';
 import { Navigate, Outlet } from 'react-router-dom';
-import { RouteList } from '../../AppRouter';
+import { RouteList } from '../../App';
 
 export const CheckNoFtOAuth = () => {
   const needFtOAuth = useAtomValue(needFtOAuthAtom);
