--- packages/client/app/src/assets/AppLogo.tsx
@@ -1,22 +1,0 @@
-import { StyleDefine } from '../styles/StyleDefine';
-import { CssPropsClassName } from '../types/CssPropsClassName';
-
-export const AppLogo = ({ className }: CssPropsClassName) => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      fill={StyleDefine.colors.textHighEmphasis}
-      viewBox="0 0 280 127"
-      className={className}
-    >
-      <title>appLogoSvg</title>
-      <path d="M192.896 112.572C190.539 112.572 188.355 112.329 186.344 111.844C184.368 111.359 182.687 110.683 181.3 109.816L183.432 103.836C184.715 104.564 186.205 105.171 187.904 105.656C189.603 106.107 191.284 106.332 192.948 106.332C194.231 106.332 195.149 106.159 195.704 105.812C196.259 105.431 196.536 104.945 196.536 104.356C196.536 103.836 196.345 103.455 195.964 103.212C195.617 102.935 195.115 102.727 194.456 102.588L188.684 101.652C186.535 101.305 184.871 100.525 183.692 99.312C182.513 98.0987 181.924 96.5213 181.924 94.58C181.924 92.7773 182.427 91.2347 183.432 89.952C184.437 88.6347 185.841 87.6293 187.644 86.936C189.447 86.208 191.509 85.844 193.832 85.844C195.669 85.844 197.489 86.0693 199.292 86.52C201.095 86.936 202.672 87.6293 204.024 88.6L201.736 94.476C200.731 93.7827 199.483 93.2107 197.992 92.76C196.501 92.2747 195.167 92.032 193.988 92.032C192.567 92.032 191.561 92.24 190.972 92.656C190.383 93.0373 190.088 93.5053 190.088 94.06C190.088 94.996 190.729 95.568 192.012 95.776L197.784 96.764C200.003 97.1107 201.701 97.8733 202.88 99.052C204.093 100.196 204.7 101.756 204.7 103.732C204.7 106.54 203.625 108.724 201.476 110.284C199.327 111.809 196.467 112.572 192.896 112.572ZM221.363 112.572C217.48 112.572 214.603 111.671 212.731 109.868C210.859 108.065 209.923 105.344 209.923 101.704V93.228H205.139V86.416H209.923V78.98H219.179V86.416H226.459V93.228H219.179V101.392C219.179 103.888 220.444 105.136 222.975 105.136C223.391 105.136 223.842 105.101 224.327 105.032C224.847 104.928 225.419 104.789 226.043 104.616L227.291 111.272C226.494 111.688 225.558 112 224.483 112.208C223.408 112.451 222.368 112.572 221.363 112.572ZM238.68 112.572C236.704 112.572 234.953 112.208 233.428 111.48C231.937 110.752 230.758 109.764 229.892 108.516C229.06 107.233 228.644 105.777 228.644 104.148C228.644 102.311 229.129 100.855 230.1 99.78C231.07 98.6707 232.63 97.8907 234.78 97.44C236.929 96.9547 239.789 96.712 243.36 96.712H245.336V95.984C245.336 94.736 245.006 93.852 244.348 93.332C243.724 92.812 242.632 92.552 241.072 92.552C239.754 92.552 238.298 92.76 236.704 93.176C235.109 93.592 233.549 94.2333 232.024 95.1L229.684 89.016C230.55 88.392 231.66 87.8373 233.012 87.352C234.364 86.8667 235.785 86.5027 237.276 86.26C238.766 85.9827 240.17 85.844 241.488 85.844C245.752 85.844 248.906 86.7627 250.952 88.6C253.032 90.4373 254.072 93.332 254.072 97.284V112H245.44V108.36C244.989 109.677 244.174 110.717 242.996 111.48C241.817 112.208 240.378 112.572 238.68 112.572ZM240.76 106.54C242.042 106.54 243.117 106.124 243.984 105.292C244.885 104.46 245.336 103.368 245.336 102.016V101.08H243.36C241.314 101.08 239.806 101.305 238.836 101.756C237.9 102.172 237.432 102.848 237.432 103.784C237.432 104.581 237.709 105.24 238.264 105.76C238.853 106.28 239.685 106.54 240.76 106.54ZM272.458 112.572C268.576 112.572 265.698 111.671 263.826 109.868C261.954 108.065 261.018 105.344 261.018 101.704V93.228H256.234V86.416H261.018V78.98H270.274V86.416H277.554V93.228H270.274V101.392C270.274 103.888 271.54 105.136 274.07 105.136C274.486 105.136 274.937 105.101 275.422 105.032C275.942 104.928 276.514 104.789 277.138 104.616L278.386 111.272C277.589 111.688 276.653 112 275.578 112.208C274.504 112.451 273.464 112.572 272.458 112.572Z" />
-      <g>
-        <path d="M99.4764 0H66.322L0 66.6717V93.6641H66.322V127H99.4764V66.6717H33.1544L99.4764 0Z" />
-        <path d="M113.678 33.3359L146.846 0H113.678V33.3359Z" />
-        <path d="M180 33.3359V0H146.846V33.3359L113.678 66.6717V100.021H146.846V66.6717L180 33.3359Z" />
-        <path d="M180 66.6717L146.845 100.021H180V66.6717Z" />
-      </g>
-    </svg>
-  );
-};
